Added 3/4 Speed and modified auton
Added the 3/4 speed due to 1/2 speed being too slow and full speed to fast. Auton was also modifed, when we got to comp we relized that when on the side with the build plate the arm would hit it and not turn properly. Changed how arm was moving to fix direction then addded a step in auton to raise and lower arm.